Remove IngestJob from IngestJobStatusTestHelper
