+++ java/ingest/ingest-core/src/test/java/sleeper/ingest/core/job/status/IngestJobStatusTestHelper.java
@@ -17,7 +17,6 @@ package sleeper.ingest.core.job.status;
 
 import sleeper.core.record.process.ProcessRunTime;
 import sleeper.core.record.process.RecordsProcessedSummary;
-import sleeper.core.record.process.status.ProcessFailedStatus;
 import sleeper.core.record.process.status.ProcessRun;
 import sleeper.core.record.process.status.ProcessRuns;
 import sleeper.core.record.process.status.TestProcessStatusUpdateRecords;
@@ -29,7 +28,6 @@ import sleeper.core.tracker.ingest.job.query.IngestJobRejectedStatus;
 import sleeper.core.tracker.ingest.job.query.IngestJobStartedStatus;
 import sleeper.core.tracker.ingest.job.query.IngestJobValidatedStatus;
 import sleeper.core.tracker.ingest.job.update.IngestJobEvent;
-import sleeper.ingest.core.job.IngestJob;
 
 import java.time.Duration;
 import java.time.Instant;
@@ -73,176 +71,6 @@ public class IngestJobStatusTestHelper {
      */
     public static IngestJobStatus ingestJobStatus(IngestJobEvent job, ProcessRun... runs) {
         return ingestJobStatus(job.getJobId(), runs);
-    }
-
-    /**
-     * Creates a process run for an ingest job that was validated.
-     *
-     * @param  job            the ingest job
-     * @param  validationTime the validation time
-     * @return                a {@link ProcessRun}
-     */
-    public static ProcessRun acceptedRun(IngestJob job, Instant validationTime) {
-        return ProcessRun.builder()
-                .startedStatus(IngestJobAcceptedStatus.from(job.getFileCount(),
-                        validationTime, defaultUpdateTime(validationTime)))
-                .build();
-    }
-
-    /**
-     * Creates a process run for an ingest job that was validated and picked up by an ingest task, but has not started
-     * yet.
-     *
-     * @param  job            the ingest job
-     * @param  taskId         the ingest task ID
-     * @param  validationTime the validation time
-     * @return                a {@link ProcessRun}
-     */
-    public static ProcessRun acceptedRunOnTask(IngestJob job, String taskId, Instant validationTime) {
-        return ProcessRun.builder()
-                .taskId(taskId)
-                .startedStatus(IngestJobAcceptedStatus.from(job.getFileCount(), validationTime,
-                        defaultUpdateTime(validationTime)))
-                .build();
-    }
-
-    /**
-     * Creates a process run for an ingest job that failed to validate.
-     *
-     * @param  job            the ingest job
-     * @param  validationTime the validation time
-     * @param  reasons        the reasons
-     * @return                a {@link ProcessRun}
-     */
-    public static ProcessRun rejectedRun(IngestJob job, Instant validationTime, String... reasons) {
-        return rejectedRun(job, null, validationTime, List.of(reasons));
-    }
-
-    /**
-     * Creates a process run for an ingest job that failed to validate.
-     *
-     * @param  job            the ingest job
-     * @param  validationTime the validation time
-     * @param  reasons        the list of reasons
-     * @return                a {@link ProcessRun}
-     */
-    public static ProcessRun rejectedRun(IngestJob job, Instant validationTime, List<String> reasons) {
-        return rejectedRun(job, null, validationTime, reasons);
-    }
-
-    /**
-     * Creates a process run for an ingest job that failed to validate.
-     *
-     * @param  job            the ingest job
-     * @param  jsonMessage    the JSON string used in ingest job deserialisation
-     * @param  validationTime the validation time
-     * @param  reasons        the list of reasons
-     * @return                a {@link ProcessRun}
-     */
-    public static ProcessRun rejectedRun(IngestJob job, String jsonMessage, Instant validationTime, List<String> reasons) {
-        return ProcessRun.builder()
-                .startedStatus(IngestJobRejectedStatus.builder()
-                        .validationTime(validationTime)
-                        .updateTime(defaultUpdateTime(validationTime))
-                        .reasons(reasons)
-                        .jsonMessage(jsonMessage)
-                        .inputFileCount(job.getFileCount())
-                        .build())
-                .build();
-    }
-
-    /**
-     * Creates a process run for an ingest job that started.
-     *
-     * @param  job       the ingest job
-     * @param  taskId    the ingest task ID
-     * @param  startTime the start time
-     * @return           a {@link ProcessRun}
-     */
-    public static ProcessRun startedIngestRun(IngestJob job, String taskId, Instant startTime) {
-        return ProcessRun.started(taskId,
-                ingestStartedStatus(job.getFileCount(), startTime, defaultUpdateTime(startTime)));
-    }
-
-    /**
-     * Creates a process run for an ingest job that finished.
-     *
-     * @param  job     the ingest job
-     * @param  taskId  the ingest task ID
-     * @param  summary the records processed summary
-     * @return         a {@link ProcessRun}
-     */
-    public static ProcessRun finishedIngestRun(
-            IngestJob job, String taskId, RecordsProcessedSummary summary) {
-        return finishedIngestRun(job, taskId, summary, 1);
-    }
-
-    /**
-     * Creates a process run for an ingest job that finished.
-     *
-     * @param  job                  the ingest job
-     * @param  taskId               the ingest task ID
-     * @param  summary              the records processed summary
-     * @param  numFilesWrittenByJob the number of files written by the job
-     * @return                      a {@link ProcessRun}
-     */
-    public static ProcessRun finishedIngestRun(
-            IngestJob job, String taskId, RecordsProcessedSummary summary, int numFilesWrittenByJob) {
-        return ProcessRun.finished(taskId,
-                ingestStartedStatus(job.getFileCount(), summary.getStartTime(), defaultUpdateTime(summary.getStartTime())),
-                IngestJobFinishedStatus.updateTimeAndSummary(defaultUpdateTime(summary.getFinishTime()), summary)
-                        .numFilesWrittenByJob(numFilesWrittenByJob).build());
-    }
-
-    /**
-     * Creates a process run for an ingest job that finished, but has not yet been committed to the state store.
-     *
-     * @param  job                  the ingest job
-     * @param  taskId               the ingest task ID
-     * @param  summary              the records processed summary
-     * @param  numFilesWrittenByJob the number of files written by the job
-     * @return                      a {@link ProcessRun}
-     */
-    public static ProcessRun finishedIngestRunUncommitted(
-            IngestJob job, String taskId, RecordsProcessedSummary summary, int numFilesWrittenByJob) {
-        return ProcessRun.finished(taskId,
-                ingestStartedStatus(job.getFileCount(), summary.getStartTime(), defaultUpdateTime(summary.getStartTime())),
-                ingestFinishedStatusUncommitted(summary, numFilesWrittenByJob));
-    }
-
-    /**
-     * Creates a process run for an ingest job that failed.
-     *
-     * @param  job            the ingest job
-     * @param  taskId         the ingest task ID
-     * @param  runTime        the process run time
-     * @param  failureReasons a list of failure reasons
-     * @return                a {@link ProcessRun}
-     */
-    public static ProcessRun failedIngestRun(
-            IngestJob job, String taskId, ProcessRunTime runTime, List<String> failureReasons) {
-        return ProcessRun.finished(taskId,
-                ingestStartedStatus(job.getFileCount(), runTime.getStartTime(), defaultUpdateTime(runTime.getStartTime())),
-                ProcessFailedStatus.timeAndReasons(defaultUpdateTime(runTime.getFinishTime()), runTime, failureReasons));
-    }
-
-    /**
-     * Creates a process run for an ingest job that passed validation then failed to start.
-     *
-     * @param  job            the ingest job
-     * @param  validationTime the validation time
-     * @param  failureTime    the failure time
-     * @param  failureReasons a list of failure reasons
-     * @return                a {@link ProcessRun}
-     */
-    public static ProcessRun acceptedAndFailedToStartIngestRun(
-            IngestJob job, Instant validationTime, Instant failureTime, List<String> failureReasons) {
-        return ProcessRun.builder()
-                .startedStatus(IngestJobAcceptedStatus.from(job.getFileCount(),
-                        validationTime, defaultUpdateTime(validationTime)))
-                .finishedStatus(ProcessFailedStatus.timeAndReasons(
-                        defaultUpdateTime(failureTime), new ProcessRunTime(failureTime, Duration.ZERO), failureReasons))
-                .build();
     }
 
     /**
